# api/openapi.yml
openapi: 3.0.0
paths:
...
  /invitations:
    post:
      operationId: InvitationsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationSummaryDto'
        '403':
          description: Caller is not an administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '409':
          description: This email already has an account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationsErrorResponseDto'
        '502':
          description: Invitation saved, but the email could not be sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationsErrorResponseDto'
      security:
        - vaultfolio_session: []
      summary: Invite a new user by email (admin only).
      tags: &ref_6
        - invitations
    get:
      operationId: InvitationsController_list
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvitationSummaryDto'
        '403':
          description: Caller is not an administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - vaultfolio_session: []
      summary: List every invitation (admin only).
      tags: *ref_6
  /invitations/{id}/cancel:
    post:
      operationId: InvitationsController_cancel
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationSummaryDto'
        '403':
          description: Caller is not an administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Invitation not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationsErrorResponseDto'
        '409':
          description: This invitation was already accepted, cancelled, or superseded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationsErrorResponseDto'
      security:
        - vaultfolio_session: []
      summary: Cancel a pending invitation (admin only).
      tags: *ref_6
  /invitations/{id}/resend:
    post:
      operationId: InvitationsController_resend
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationSummaryDto'
        '403':
          description: Caller is not an administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: Invitation not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationsErrorResponseDto'
        '409':
          description: This invitation was already accepted, cancelled, or superseded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationsErrorResponseDto'
        '502':
          description: Invitation could not be re-sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationsErrorResponseDto'
      security:
        - vaultfolio_session: []
      summary: Resend a pending invitation email (admin only).
      tags: *ref_6
  /invitations/token/{token}:
    get:
      operationId: InvitationsController_lookupToken
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationTokenLookupDto'
        '410':
          description: This invitation link is no longer valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationsErrorResponseDto'
      summary: Look up a pending invitation's email/role by its token.
      tags: *ref_6
  /invitations/token/{token}/accept:
    post:
      operationId: InvitationsController_accept
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInvitationRequestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionUserDto'
        '400':
          description: Password must be between 8 and 200 characters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationsErrorResponseDto'
        '410':
          description: This invitation link is no longer valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationsErrorResponseDto'
      summary: Accept an invitation, creating the account and starting a session.
      tags: *ref_6
...
info:
  title: Vaultfolio API
  description: >-
    Generated directly from the backend controllers/DTOs — see
    specs/031-openapi-swagger-integration. "Try it out" reuses the
    vaultfolio_session session cookie from a normal app login; it is not a
    separate credential scheme.
  version: 0.0.9
  contact: {}
...
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: vaultfolio_session
  schemas:
...
    SessionUserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: admin@example.com
        displayName:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
        domainScopes:
          description: Domain ids this user is entitled to, independent of `role`.
          type: array
          items:
            type: string
      required:
        - id
        - email
        - displayName
        - role
        - domainScopes
...
    ErrorResponseDetailDto:
      type: object
      properties:
        field:
          type: string
          example: quantity
        message:
          type: string
          example: Must be a positive number.
      required:
        - field
        - message
    ErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          description: Stable machine-readable error code.
          example: not_found
        message:
          type: string
          example: Account not found.
        correlationId:
          type: string
          description: >-
            UUID for end-to-end tracing; matches the X-Correlation-Id response
            header.
          format: uuid
        details:
          description: >-
            Only present for field-level validation errors; non-empty when
            present.
          type: array
          items:
            $ref: '#/components/schemas/ErrorResponseDetailDto'
      required:
        - error
        - message
        - correlationId
...
    CreateInvitationRequestDto:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
      required:
        - email
        - role
    InvitationSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
        status:
          type: string
          enum:
            - PENDING
            - ACCEPTED
            - EXPIRED
            - CANCELLED
            - SUPERSEDED
        invitedBy:
          type: string
          format: uuid
          description: Id of the admin who sent this invitation.
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - status
        - invitedBy
        - createdAt
        - expiresAt
    InvitationsErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          enum:
            - account_exists
            - not_found
            - already_resolved
            - email_delivery_failed
            - invalid_invitation
            - invalid_password
        message:
          type: string
      required:
        - error
        - message
    InvitationTokenLookupDto:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
      required:
        - email
        - role
    AcceptInvitationRequestDto:
      type: object
      properties:
        password:
          type: string
          format: password
        displayName:
          type: string
      required:
        - password
        - displayName
...